Enable multistream reader to read from truncated
When a projection reads from multiple streams, the multistream reader
currently assumes that a read will either return events or not. In cases
where it doesn't return any events, it will assume that it's the end of
a stream and it will not update the position it reads from for that
stream.

When a stream is truncated at say event number 10, when a read from 0
with a count of 1 event completes, the read will return a next position
of 10 and no events. This doesn't mean that the stream is at it's end,
it just means that the stream has been truncated and that the stream
starts at position 10.

Reproduction steps
1. Create a stream with 10 events.
2. Write 1 event to the stream.
3. Create a projection that reads from multiple streams including this
one
4. The projection will not process any events.